requirements.txt
APScheduler==3.11.2
feedparser==6.0.12
Flask==3.0.3
groq==1.1.2
gunicorn==22.0.0
pillow==12.1.1
python-dotenv==1.0.1
pytz==2026.1.post1
requests==2.32.3
